+++ packages/i18n/CHANGELOG.md
@@ -2,6 +2,66 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.0.1-alpha.69](https://github.com/GetStream/stream-video-js/compare/@stream-io/i18n-0.0.1-alpha.68...@stream-io/i18n-0.0.1-alpha.69) (2023-05-26)
+
+
+
+## [0.0.1-alpha.68](https://github.com/GetStream/stream-video-js/compare/@stream-io/i18n-0.0.1-alpha.67...@stream-io/i18n-0.0.1-alpha.68) (2023-05-26)
+
+
+
+## [0.0.1-alpha.67](https://github.com/GetStream/stream-video-js/compare/@stream-io/i18n-0.0.1-alpha.66...@stream-io/i18n-0.0.1-alpha.67) (2023-05-26)
+
+
+
+## [0.0.1-alpha.66](https://github.com/GetStream/stream-video-js/compare/@stream-io/i18n-0.0.1-alpha.65...@stream-io/i18n-0.0.1-alpha.66) (2023-05-25)
+
+
+
+## [0.0.1-alpha.65](https://github.com/GetStream/stream-video-js/compare/@stream-io/i18n-0.0.1-alpha.64...@stream-io/i18n-0.0.1-alpha.65) (2023-05-25)
+
+
+
+## [0.0.1-alpha.64](https://github.com/GetStream/stream-video-js/compare/@stream-io/i18n-0.0.1-alpha.63...@stream-io/i18n-0.0.1-alpha.64) (2023-05-25)
+
+
+
+## [0.0.1-alpha.63](https://github.com/GetStream/stream-video-js/compare/@stream-io/i18n-0.0.1-alpha.62...@stream-io/i18n-0.0.1-alpha.63) (2023-05-25)
+
+
+
+## [0.0.1-alpha.62](https://github.com/GetStream/stream-video-js/compare/@stream-io/i18n-0.0.1-alpha.61...@stream-io/i18n-0.0.1-alpha.62) (2023-05-25)
+
+
+
+## [0.0.1-alpha.61](https://github.com/GetStream/stream-video-js/compare/@stream-io/i18n-0.0.1-alpha.60...@stream-io/i18n-0.0.1-alpha.61) (2023-05-25)
+
+
+
+## [0.0.1-alpha.60](https://github.com/GetStream/stream-video-js/compare/@stream-io/i18n-0.0.1-alpha.59...@stream-io/i18n-0.0.1-alpha.60) (2023-05-25)
+
+
+
+## [0.0.1-alpha.59](https://github.com/GetStream/stream-video-js/compare/@stream-io/i18n-0.0.1-alpha.58...@stream-io/i18n-0.0.1-alpha.59) (2023-05-25)
+
+
+
+## [0.0.1-alpha.58](https://github.com/GetStream/stream-video-js/compare/@stream-io/i18n-0.0.1-alpha.57...@stream-io/i18n-0.0.1-alpha.58) (2023-05-25)
+
+
+
+## [0.0.1-alpha.57](https://github.com/GetStream/stream-video-js/compare/@stream-io/i18n-0.0.1-alpha.56...@stream-io/i18n-0.0.1-alpha.57) (2023-05-24)
+
+
+
+## [0.0.1-alpha.56](https://github.com/GetStream/stream-video-js/compare/@stream-io/i18n-0.0.1-alpha.55...@stream-io/i18n-0.0.1-alpha.56) (2023-05-24)
+
+
+
+## [0.0.1-alpha.55](https://github.com/GetStream/stream-video-js/compare/@stream-io/i18n-0.0.1-alpha.54...@stream-io/i18n-0.0.1-alpha.55) (2023-05-24)
+
+
+
 ## [0.0.1-alpha.54](https://github.com/GetStream/stream-video-js/compare/@stream-io/i18n-0.0.1-alpha.53...@stream-io/i18n-0.0.1-alpha.54) (2023-05-23)